+++ tests/testthat/test_set_censobr_cache_dir.R
@@ -11,14 +11,17 @@ testthat::skip_if_not_installed("arrow")
 test_that("set_censobr_cache_dir", {
 
   # default
-  testthat::expect_message( set_censobr_cache_dir() )
+  testthat::expect_message( set_censobr_cache_dir(path = NULL) )
 
   # Set custom cache directory
   tempd <- tempdir()
   testthat::expect_message( set_censobr_cache_dir(path = tempd) )
 
+  current_dir <- get_censobr_cache_dir()
+  testthat::expect_true(identical(basename(current_dir), basename(tempd)))
+
   # download
-  t <- read_emigration(year = 2010, showProgress = FALSE)
+  read_emigration(year = 2010, showProgress = FALSE)
 
   # check if file exists in custom dir
   files <- list.files(tempd, full.names = TRUE)
@@ -34,8 +37,7 @@ test_that("set_censobr_cache_dir", {
 # ERRORS and messages  -----------------------
 test_that("set_censobr_cache_dir", {
 
-  # Wrong date 4 digits
   testthat::expect_error(set_censobr_cache_dir(path = 999))
-  testthat::expect_error(set_censobr_cache_dir(path = '999'))
+
 })