+++ tests/testthat/test_z_censobr_cache.R
@@ -10,17 +10,18 @@ testthat::skip_if_not_installed("arrow")
 
 test_that("censobr_cache", {
 
-  # simply list files
-  testthat::expect_message( censobr_cache() )
+  # clean cache
+  censobr_cache(delete_file = 'all')
 
-  ## delete existing
+  # simply list files
+  testthat::expect_message( censobr_cache(list_files = TRUE) )
+
+  # default cache dir
+  censobr::set_censobr_cache_dir(path = NULL)
+  cache_dir <- censobr::get_censobr_cache_dir()
 
   # download
   censobr::read_emigration(year = 2010, showProgress = FALSE, cache = TRUE)
-
-  # cache dir
-  pkgv <- paste0('censobr/data_release_', censobr_env$data_release)
-  cache_dir <- tools::R_user_dir(pkgv, which = 'cache')
 
   # list cached files
   files <- list.files(cache_dir, full.names = TRUE)
@@ -39,10 +40,20 @@ test_that("censobr_cache", {
   testthat::expect_true( file.exists(fname_full) )
   testthat::expect_message( censobr_cache(delete_file = 'all') )
   censobr_cache(delete_file = 'all')
-#  testthat::expect_true( length(list.files(cache_dir)) == 0 )
+  testthat::expect_true( length(list.files(cache_dir)) == 0 )
 
   # if file does not exist, simply print message
   testthat::expect_message( censobr_cache(delete_file ='aaa') )
+
+
+  # check is output is "fs_path" when print_tree = TRUE
+  censobr::read_emigration(year = 2010, showProgress = FALSE, cache = TRUE)
+  testthat::expect_message(censobr_cache(list_files = TRUE, print_tree = TRUE))
+  temp <- censobr_cache(list_files = TRUE, print_tree = TRUE)
+  testthat::expect_is(temp, "fs_path")
+
+  # get current cache dir
+  testthat::expect_true(is.character(get_censobr_cache_dir()))
 
  })
 
@@ -52,7 +63,12 @@ test_that("censobr_cache", {
 
   # Wrong date 4 digits
   testthat::expect_error(censobr_cache(list_files= 999))
+  testthat::expect_error(censobr_cache(print_tree = 999))
   testthat::expect_error(censobr_cache(delete_file = 999))
+
+  # check is output is "fs_path" when print_tree = TRUE
+  testthat::expect_error(censobr_cache(list_files = FALSE, print_tree = TRUE))
+
   })